Clearify role of atomic form factors
There seems to be a confusion for many people when talking about
form factors. In some communities the term
form factor is used for the scattering intensity. To reudce the
confusion, I added the prefix 'atomic' when we
talk about the form factor and added a note box to the docstrings functions.